add model

+++ DoAnTotNghiep/Entity/City.cs
@@ -9,14 +9,18 @@ namespace DoAnTotNghiep.Entity
         [Key]
         [Column("id")]
         public int Id { get; set; }
+
         [Required(ErrorMessage = "Hãy điền tên thành phố")]
         [MaxLength(100, ErrorMessage = "Tên thành phố tối đa 100 ký tự")]
         [Column("name")]
         public string Name { get; set; } = string.Empty;
+
         [Column("lat")]
-        public float Lat { get; set; } = 0;
+        public double Lat { get; set; } = 0;
+
         [Column("long")]
-        public float Long { get; set; } = 0;
+        public double Long { get; set; } = 0;
+
         [Column("count")]
         public int Count { get; set; } = 0;
     }
